+++ mockserver/mockserver-core/src/main/java/org/mockserver/model/Body.java
@@ -61,6 +61,7 @@ public abstract class Body<T> extends Not {
         JSON_PATH,
         PARAMETERS,
         REGEX,
+        FUZZY,
         STRING,
         XML,
         XML_SCHEMA,